Made Serial interrupt driven
I have removed the delay issue from the stepper code and have made it interrupt driven.  If at anytime the # character, followed by enter, is received it will stop the stepper motor.  Not sure how we could stop the servos.  I guess we could use a mosfet to turn on/off the power to the servos, allowing them to become unpowered if something goes wrong.  Maybe something to add to the schematic and mention in the report but im not sure if we can implement that in time.
I still need to implement the ability to send over multiple motor/joint commands in one message like how we did for gantry system.
I also updated to use a RX and TX buffer (char array) rather than a "string"